% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_additional_functions.r
\name{JAGS.objects}
\alias{JAGS.objects}
\title{JAGS output an internal function used within the tuts package}
\usage{
JAGS.objects(JAGS.output)
}
\arguments{
\item{JAGS.output}{An JAGS MCMC Object.}
}
\description{
\code{JAGS.objects} an internal function used within the tuts package.
}
