% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuar1.R
\name{plot.tuts_ar1}
\alias{plot.tuts_ar1}
\title{Plots and visual diagnostics of tuts_ar1 objects}
\usage{
\method{plot}{tuts_ar1}(x, type, ...)
}
\arguments{
\item{x}{A tuts_ar1 object.}

\item{type}{plot type with the following options:\cr
- 'predTUTS' plots one step predictions of the model. \cr
- 'par' plots distributions of parameters of the model. \cr
- 'volatility' plots volatility realizations. \cr
- 'GR' plots Gelman-Rubin diagnostics. \cr
- 'cv' plots 5-fold cross validation. \cr
- 'mcmc' plots diagnostics of MCMC/JAGS objects. \cr}

\item{...}{list of optional parameters:\cr
- burn: burn-in parameter ranging from 0 to 0.7 with default value set to 0. \cr
- CI: credible interval ranging from 0.3 to 1 with default value set to 0.95.}
}
\description{
\code{plot.tuts_ar1(x,type,...)} generates plots and visual diagnostics of tuts_ar1 objects.
}
\examples{
# Note: Most of models included in tuts package are computationally intensive. In the example
# below I set parameters to meet CRAN’s testing requirement of maximum 5 sec per example.
# A more practical example would contain N=50 in the first line of the code and n.sim=10000.

#1. Import or simulate the data (a simulation is chosen for illustrative purposes):
DATA=simtuts(N=10,Harmonics=c(4,0,0), sin.ampl=c(10,0, 0), cos.ampl=c(0,0,0),
            trend=0,y.sd=2, ti.sd=0.2)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(0.2, length(ti.mu))

#2. Fit the model:
n.sim=1000
TUAR1=tuar1(y=y,ti.mu=ti.mu,ti.sd=ti.sd,n.sim=n.sim,CV=TRUE,n.cores=2)

#3. Generate plots and diagnostics of the model (optional parameters are listed in brackets):
plot(TUAR1,type='predTUTS')               # One step out of salmple predictions (CI, burn).
plot(TUAR1,type='par', burn=0.4)          # Distributions of parameters (burn).
plot(TUAR1,type='mcmc')                   # MCMC diagnostics.
plot(TUAR1,type='cv', burn=0.4, CI=0.9)   # 5 fold cross validation (CI, burn).
plot(TUAR1,type='GR')                     # Gelman-Rubin diagnostic (CI, burn).
plot(TUAR1,type='volatility')             # Volatility realizaitons.
}
