% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuar1redf.R
\name{plot.tuts_ar1redf}
\alias{plot.tuts_ar1redf}
\title{Graphical summaries and diagnostics of tuts_ar1redf objects}
\usage{
\method{plot}{tuts_ar1redf}(x, type, ...)
}
\arguments{
\item{x}{A tuts_tuar1 objects.}

\item{type}{plot/disgnostics type (options:
'predTUTS' plots one step preditions of the model, 'GR' plots Gelman-Rubin diagnostics,
'cv' plots 5-fold cross validation, 'mcmc' plots diagnostics of mcmc objects)}

\item{...}{list of optional parameters: 'burn' (burn-in parameter ranging from 0 to 0.7 with
default value set to 0), and CI (credible interval ranging from 0.3 to 1 with
default value set to 0.95)}
}
\description{
\code{plot.tuts_ar1redf(x,type,...)} plots summaries and diagnostics of tuts_ar1redf objects.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the tuar1redf model with cross validation:
n.sim=1000; n.chains=2
AR1REDF=tuar1redf(y=y,ti.mu=ti.mu,ti.sd=ti.sd,n.sim=n.sim,n.chains=n.chains, CV=TRUE)

# Plots and diagnostics (optional parameters are listed in brackets):
plot(AR1REDF,type='predTUTS',burn=0.2,CI=0.99)  # One step out of salmple predictions of the model (CI, burn).
plot(AR1REDF,type='par', burn=0.4)              # Distributions of parameters; (burn).
plot(AR1REDF,type='mcmc')                       # mcmc diagnostics.
plot(AR1REDF,type='cv', burn=0.4)               # 5 fold cross validation plot (CI, burn).
plot(AR1REDF,type='GR', burn=0.4)               # Gelman-Rubin diagnostic (CI, burn).
plot(AR1REDF,type='tau')                        # Volatility realizaitons.

}
