% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tubfs.R
\name{plot.tuts_BFS}
\alias{plot.tuts_BFS}
\title{Graphical summaries and diagnostics of tuts_BFS objects.}
\usage{
\method{plot}{tuts_BFS}(x, type, ...)
}
\arguments{
\item{x}{A tuts_BFS objects.}

\item{type}{plot/disgnostic type (options: 'periodogram' plots estimates of power spectrum,
'predTUTS' plots one step predictions of the model, 'GR' plots Gelman-Rubin diagnostics,
'cv' plots 5-fold cross validation, 'mcmc' plots diagnostics of mcmc objects, and 'volatility' plots
 volatility realizations)}

\item{...}{list of optional parameters: 'burn' (burn-in parameter ranging from 0 to 0.7 with
default value set to 0), and CI (credible interval ranging from 0.3 to 1 with
default value set to 0.95)}
}
\description{
\code{plot.tuts_BFS(x,...)} plots and diagnostics of tuts_BFS objects.
}
\examples{
# Generate actual and observed time series as a sum of 2 sine functions:
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the Bayesian Frequency Selection model with cross validation of the model:
BFS=tubfs(y=y,ti.mu=ti.mu,ti.sd=ti.sd,freqs='internal',n.sim=n.sim,n.chains=2, CV=TRUE)

# Plots and diagnostics:
plot(BFS,type='periodogram')               # spectral analysis; optional parameters: CI, burn.
plot(BFS,type='predTUTS', CI=0.99)         # One step predictions of the model; optional: CI, burn.
plot(BFS,type='cv')                        # 5 fold cross validation plot; optional: CI, burn.
plot(BFS,type='GR')                        # Gelman-Rubin diagnostics; optional: CI, burn.
plot(BFS,type='mcmc')                      # mcmc diagnostics.
plot(BFS,type='volatility')                # Volatility realizaitons.
}
