% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_tupoisPN.R
\name{plot.tuts_poisPN}
\alias{plot.tuts_poisPN}
\title{Graphical summaries and diagnostics of tuts_poisPN objects}
\usage{
\method{plot}{tuts_poisPN}(x, type, ...)
}
\arguments{
\item{x}{A tuts_poisPN object.}

\item{type}{plot/diagnostic type (options:'predTUTS' plots one step out of sample predictions of the model,
'GR' plots Gelman-Rubin diagnostics, 'cv' plots 5-fold cross validation, 'mcmc' plots diagnostics
 of mcmc objects, and 'lambda' plots lambda realizations).}

\item{...}{list of optional parameters: 'burn' (burn-in parameter ranging from 0 to 0.7 with
default value set to 0), and CI (credible interval ranging from 0.3 to 1 with
default value set to 0.95).}
}
\description{
\code{plot.tuts_poisPN} plots summaries and diagnostics of a tuts_poisPN object.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0), trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
y=round(y-min(y))
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Set parameters and run the polynomial regression of counting series:
polyorder=2
n.sim=1000
PPN=tupoispn(y=y,ti.mu=ti.mu,ti.sd=ti.sd,polyorder=polyorder,n.sim=n.sim, CV=TRUE)

# Plots and diagnostics (optional parameters are listed in brackets):
plot(PPN,type='predTUTS',CI=0.95)      # One step out of salmple predictions of the model (CI, burn).
plot(PPN,type='cv',burn=0.3)           # 5 fold cross-validation (CI, burn).
plot(PPN,type='GR',CI=0.95)            # Gelman-Rubin diagnostic (CI).
plot(PPN,type='mcmc')                  # MCMC diagnostics.
plot(PPN,type='lambda')                # Volatility realizaitons.

}
