% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuls.r
\name{summary.tuts_ls}
\alias{summary.tuts_ls}
\title{Function returning a list of frequencies having significant power estimates.}
\usage{
\method{summary}{tuts_ls}(x)
}
\arguments{
\item{x}{A tuts_LS obect.}
}
\description{
\code{summary.tuts_LS} Function returning a list of frequencies for which spectral power exceeds prespecified
 confidence interval contained in tuts_LS objects.
}
\examples{
# Simulate time-uncertain time series:
DATA=simtuts(N=100,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)

# (Set parameters and ) run multiple Lomb-Scargle periodograms:
TULS=tuls(y=y,ti.mu=ti.mu,ti.sd=ti.sd)       # Optional parameters: ofac, CI, n.sim.

# Obtain list of frequencies for which spectral power exceeds prespecified confidence interval
summary(TULS)
}
