% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_additional_functions.r
\name{JAGS.objects}
\alias{JAGS.objects}
\title{JAGS output}
\usage{
JAGS.objects(JAGS.output)
}
\arguments{
\item{JAGS.output}{An JAGS MCMC Object.}
}
\description{
\code{JAGS.objects(JAGS.output)} an internal function used within the tuts package. Additional
 plots and diagnostics are available from model-specific methods, for more information see the
 help documents of individual models.
}
