% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuar1.R
\name{summary.tuts_ar1}
\alias{summary.tuts_ar1}
\title{Prints summary of tuts_ar1 objects}
\usage{
\method{summary}{tuts_ar1}(x, ...)
}
\arguments{
\item{x}{A tuts_ar1 object.}

\item{...}{list of optional parameters. The list contains burn-in parameter
ranging from 0 to 0.5, with the default value burn=0, and the credible interval parameter
ranging between 0.5 and 1, with the default CI=0.99.}
}
\description{
\code{summary.tuts_ar1(x,...)} Prints summary of a tuts_ar1 object.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the tuar1 model fitting with cross validation:
n.sim=1000
TUAR1=tuar1(y=y,ti.mu=ti.mu,ti.sd=ti.sd,n.sim=n.sim, CV=TRUE)

# Generate summary results (optional parameters are listed in brackets):
summary(TUAR1)                                # Summary tables (CI, burn).

}
