% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuar1redf.R
\name{summary.tuts_ar1redf}
\alias{summary.tuts_ar1redf}
\title{Summaries of tuts_ar1redf objects}
\usage{
\method{summary}{tuts_ar1redf}(x, ...)
}
\arguments{
\item{object}{A tuts_ar1redf object.}
}
\description{
\code{summary.tuts_ar1redf(x,...)} Prints summary of a tuts_ar1redf object.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the tuar1redf model with cross validation:
n.sim=1000; n.chains=2
AR1REDF=tuar1redf(y=y,ti.mu=ti.mu,ti.sd=ti.sd,n.sim=n.sim,n.chains=n.chains, CV=TRUE)

# Generate summary results (optional parameters are listed in brackets):
summary(AR1REDF)                                # Summary statistics (urn, CI).
}
