% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tubfs.R
\name{summary.tuts_BFS}
\alias{summary.tuts_BFS}
\title{Summary of tuts_BFS objects.}
\usage{
\method{summary}{tuts_BFS}(x, ...)
}
\arguments{
\item{x}{A tuts_BFS object.}

\item{...}{A list of optional parameters.burn: burn-in parameter
ranging from 0 to 0.7, the default value is 0.CI: confidence interval, default is set to 0.99.}
}
\description{
\code{summary.tuts_BFS(x,...)} prints summaries of tuts_BFS objects.
}
\examples{
# Generate actual and observed time series as a sum of 2 sine functions:
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the Bayesian Frequency Selection model with cross validation of the model:
BFS=tubfs(y=y,ti.mu=ti.mu,ti.sd=ti.sd,freqs='internal',n.sim=n.sim,n.chains=2, CV=TRUE)

# Generate summary results:
summary(BFS)                               # Optional parameters: CI, burn.
summary(BFS,burn=0.2)                      # Results after 20\% of burn-in.

}
