% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_tupoisBFS.R
\name{summary.tuts_poisBFS}
\alias{summary.tuts_poisBFS}
\title{Summary of tuts_poisBFS objects.}
\usage{
\method{summary}{tuts_poisBFS}(x, ...)
}
\arguments{
\item{x}{A tuts_poisBFS object.}

\item{...}{A list of optional parameters.burn: burn-in parameter
ranging from 0 to 0.7, the default value is 0.CI: confidence interval, default is set to 0.99.}
}
\description{
\code{summary.tuts_poisBFS(x,...)} prints summaries of tuts_poisBFS objects.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs ; y=round(y)-min(round(y))
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the Bayesian Frequency Selection model with cross validation of the model:
n.sim=1000
TUPOIS=tupoisbsf(y=y,ti.mu=ti.mu,ti.sd=ti.sd,freqs='internal',n.sim=n.sim,n.chains=2, CV=TRUE)

# Generate summary results (optional parameters are listed in brackets):
summary(TUPOIS)                               # Summary results (CI, burn).
summary(TUPOIS,burn=0.2)                      # Results after 20\% of burn-in (CI).

}
