% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuPN.R
\name{summary.tuts_polyn}
\alias{summary.tuts_polyn}
\title{Prints summaries of tuts_polyn objects.}
\usage{
\method{summary}{tuts_polyn}(x, ...)
}
\arguments{
\item{x}{A tuts_polyn object.}

\item{...}{list of optional parameters. The list contains burn-in parameter
ranging from 0 to 0.5, with the default value burn=0, and the credible interval parameter
ranging between 0.5 and 1, with the default CI=0.99.}
}
\description{
\code{summary.tuts_polyn} Prints summaries of tuts_polyn objects.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0), trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Set parameters and run the polynomial regression:
polyorder=2
n.sim=1000
PN=tupolyn(y=y,ti.mu=ti.mu,ti.sd=ti.sd,polyorder=polyorder,n.sim=n.sim, CV=TRUE)

# Generate summary results (optional parameters are listed in brackets):
summary(PN)                               # Summary statistics (burn, CI).

}
