% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_additional_functions.r
\name{summary.tuts_wrap}
\alias{summary.tuts_wrap}
\title{Summary table comparing models based on DIC criterion.}
\usage{
\method{summary}{tuts_wrap}(x)
}
\description{
\code{summary.tuts_wrap(x)} Summary table of a tuts_wrap objects.
}
