% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuar1.R
\name{tuar1}
\alias{tuar1}
\title{Time-uncertain AR(1) model}
\usage{
tuar1(y, ti.mu, ti.sd, n.sim, CV = FALSE, ...)
}
\arguments{
\item{y}{A vector of observations.}

\item{ti.mu}{A vector of estimates of timing of observations.}

\item{ti.sd}{A vector of standard deviations of timing.}

\item{n.sim}{A number of simulations.}

\item{CV}{cross-validation indicator.}

\item{...}{A list of optional parameters. The list contains thinning parameter,
with the default value set to Thin=4, and the number of mcmc chains with the default value set
 to n.chains=2.}
}
\description{
\code{tuar1} estimates unbiased parameters of AR(1) model.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the tuar1 model fitting with cross validation:
n.sim=1000
TUAR1=tuar1(y=y,ti.mu=ti.mu,ti.sd=ti.sd,n.sim=n.sim, CV=TRUE)

# Generate summary results (optional parameters are listed in brackets):
summary(TUAR1)                                # Summary tables (CI, burn).

# Plots and diagnostics (optional parameters are listed in brackets):
plot(TUAR1,type='predTUTS')                   # One step out of salmple predictions of the model (CI, burn).
plot(TUAR1,type='par', burn=0.4)              # Distributions of parameters of the AR(1) model (burn).
plot(TUAR1,type='mcmc')                       # mcmc diagnostics.
plot(TUAR1,type='cv', burn=0.4, CI=0.9)       # 5 fold cross validation plot(CI, burn).
plot(TUAR1,type='GR')                         # Gelman-Rubin diagnostic (CI, burn).
plot(TUAR1,type='volatility')                 # Volatility realizaitons.

}
