% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tuls.r
\name{tuls}
\alias{tuls}
\title{Spectral analysis of time-uncertain time series using Lomb-Scargle periodogram}
\usage{
tuls(y, ti.mu, ti.sd, ...)
}
\arguments{
\item{y}{A vector of observations.}

\item{ti.mu}{A vector of estimates of timings of observations.}

\item{ti.sd}{A vector of standard deviations of timings.}

\item{...}{A list of optional parameters. The list contains thinning oversamlping parameter,
the default value is ofac=4, confidence interval, the default value is CI=0.99, and the number
 of simulated timings, wiith the default vale set to n.sim=1000.}

\item{n.sim}{A number of simulations.}
}
\description{
\code{tuls} computes multiple power estimates using the Lomb-Scargle algorithm and simulated realizations of
timings of observations. Timings are simulated with normal distribution \emph{ti~N(ti.mu,ti.sd)},
and sorted in ascending order to ensure non-overlapping feature of observations.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# (Set parameters and ) run multiple Lomb-Scargle periodograms:
TULS=tuls(y=y,ti.mu=ti.mu,ti.sd=ti.sd)       # Optional parameters: ofac, CI, n.sim

# Plot the results:
plot(TULS)

# Obtain list of frequencies for which spectral power exceeds confidence interval prespecified in the tuts_ls class
summary(TULS)

}
\references{
\url{https://en.wikipedia.org/wiki/Least-squares_spectral_analysis}
}
\seealso{
\url{http://cran.r-project.org/package=Bchron}
}
