% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_tupoisBFS.R
\name{tupoisbsf}
\alias{tupoisbsf}
\title{Time uncertain Poisson regression with the Bayesian Frequency Selection method}
\usage{
tupoisbsf(y, ti.mu, ti.sd, n.sim, CV = FALSE, ...)
}
\arguments{
\item{y}{A vector of observations.}

\item{ti.mu}{A vector of estimates/observed timings of observations.}

\item{ti.sd}{A vector of standard deviations of timings.}

\item{n.sim}{A number of simulations.}

\item{CV}{cross-validation indicator.}

\item{...}{optional arguments. n.chains: number of MCMC chains, default is set to 2.
Thin: Thinning factor, default is set to 4.m: maximum number of significant frequencies
 in the data, the default is set to 5. polyorder: order of the polynomial regression component,
 the default is set to 3. freqs: set to a positive integer k returns a vector of k equally spaced frequencies in the Nyquist
  range. freqs can be provided as a vector of custom frequencies of interest. Set to 'internal' generates an equally
  spaced vector of frequencies in the Nyquist range}
}
\description{
\code{tupoisbsf}
spectral analysis of time-uncertain time series of count data sets using the Bayesian Frequency Selection method
 described in the paper \emph{"Frequency selection in palaeoclimate time series: a model-based
  approach incorporating possible time uncertainty"} by P. Franke, Prof B. Huntley, Dr A. Parnell.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0),
trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs ; y=round(y)-min(round(y))
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Run the Bayesian Frequency Selection model with cross validation of the model:
n.sim=1000
TUPOIS=tupoisbsf(y=y,ti.mu=ti.mu,ti.sd=ti.sd,freqs='internal',n.sim=n.sim,n.chains=2, CV=TRUE)

# Generate summary results (optional parameters are listed in brackets):
summary(TUPOIS)                               # Summary results (CI, burn).
summary(TUPOIS,burn=0.2)                      # Results after 20\% of burn-in (CI).

# Plots and diagnostics (optional parameters are listed in brackets):
plot(TUPOIS,type='periodogram')               # spectral analysis (CI, burn).
plot(TUPOIS,type='predTUTS', CI=0.99)         # One step predictions of the model (CI, burn).
plot(TUPOIS,type='cv')                        # 5 fold cross validation plot (CI, burn).
plot(TUPOIS,type='GR')                        # Gelman-Rubin diagnostics (CI, burn).
plot(TUPOIS,type='mcmc')                      # mcmc diagnostics.
plot(TUPOIS,type='lambda')                    # Realizaitons of lambda.

}
