% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_tupoisPN.R
\name{tupoispn}
\alias{tupoispn}
\title{Time-uncertain polynomial regression of counting series}
\usage{
tupoispn(y, ti.mu, ti.sd, n.sim, CV = FALSE, ...)
}
\arguments{
\item{y}{A vector of observations.}

\item{ti.mu}{A vector of estimates of timing of observations.}

\item{ti.sd}{A vector of standard deviations of timing.}

\item{n.sim}{A number of simulations.}

\item{CV}{cross-validation indicator.}

\item{...}{A list of optional parameters. The list contains polynomial order with the default value
set to polyorder=3, thinning parameter, with the default value set to Thin=4,
and the number of mcmc chains with the default value set to n.chains=2.}

\item{polyorder}{Order of the polynomial regression.}
}
\description{
\code{tupoispn}  poisson polynomial regression of the N-th order of time-uncertain time series.
}
\examples{
# Import or simulate the data (simulation is chosen for illustrative purposes):
DATA=simtuts(N=50,Harmonics=c(10,30,0), sin.ampl=c(10,10, 0), cos.ampl=c(0,0,0), trend=0,y.sd=3, ti.sd=1)
y=DATA$observed$y.obs
y=round(y-min(y))
ti.mu=DATA$observed$ti.obs.tnorm
ti.sd= rep(1, length(ti.mu))

# Set parameters and run the polynomial regression of counting series:
polyorder=2
n.sim=1000
PPN=tupoispn(y=y,ti.mu=ti.mu,ti.sd=ti.sd,polyorder=polyorder,n.sim=n.sim, CV=TRUE)

# Generate summary results (optional parameters are listed in brackets):
summary(PPN)                           # Summary statistics (burn, CI).

# Plots and diagnostics (optional parameters are listed in brackets):
plot(PPN,type='predTUTS',CI=0.95)      # One step out of salmple predictions of the model (CI, burn).
plot(PPN,type='cv',burn=0.3)           # 5 fold cross-validation (CI, burn).
plot(PPN,type='GR',CI=0.95)            # Gelman-Rubin diagnostic (CI).
plot(PPN,type='mcmc')                  # MCMC diagnostics.
plot(PPN,type='lambda')                # Volatility realizaitons.

}
